Multi class kernel based crammer singer SVM
This project is implementation of the multi class SVM Crammer Singer which uses Karush-Kuhn-Tucker conditions to choose sample in each iteration of optimization in python code. Until now, despite OVR SVM, there is no library for kernel based multi class SVM in crammer singer method. To run optimally I chose a subset of 5000 samples from MNIST training set and a subset of 1000 samples form testing set. The implementation can be done with or without kernel functions to reach the optimum hyperplane in classification. We do not implement stuff related to speed up the training process like using Active Set, using cooling strategy on epsilon and caching samples in this project.
This project is a class project in Pattern Recognition course with a focus on support vector machines. The explanation of the project will be posted soon.